requirements.txt
fastapi
uvicorn
flask
sqlalchemy
psycopg2-binary
requests
python-dotenv
email-validator
Pillow
aiofiles
python-multipart
bcrypt
python-jose[cryptography]
